Fix npc_text for classic.

+++ WowPacketParser/Parsing/Parsers/QuestHandler.cs
@@ -960,12 +960,12 @@ namespace WowPacketParser.Parsing.Parsers
             if ((flags & 0x3) == 0) // not completable
             {
                 requestItems.EmoteOnIncompleteDelay = emoteDelay;
-                requestItems.EmoteOnIncomplete = emoteID;
+                requestItems.EmoteOnIncomplete = (int)emoteID;
             }
             else
             {
                 requestItems.EmoteOnCompleteDelay = emoteDelay;
-                requestItems.EmoteOnComplete = emoteID;
+                requestItems.EmoteOnComplete = (int)emoteID;
             }
             packet.ReadUInt32("Unk flags 2");
             packet.ReadUInt32("Unk flags 3");
@@ -996,7 +996,7 @@ namespace WowPacketParser.Parsing.Parsers
             };
 
             requestItems.EmoteOnCompleteDelay = packet.ReadUInt32("Delay");  // not confirmed
-            requestItems.EmoteOnComplete = packet.ReadUInt32("Emote");  // not confirmed
+            requestItems.EmoteOnComplete = (int)packet.ReadUInt32("Emote");  // not confirmed
 
             packet.ReadUInt32("Close Window on Cancel");
             packet.ReadUInt32E<QuestFlags>("Quest Flags");
@@ -1042,7 +1042,7 @@ namespace WowPacketParser.Parsing.Parsers
                 CompletionText = text
             };
 
-            requestItems.EmoteOnComplete = packet.ReadUInt32("Emote");
+            requestItems.EmoteOnComplete = (int)packet.ReadUInt32("Emote");
             requestItems.EmoteOnCompleteDelay = packet.ReadUInt32("Delay");
             packet.ReadUInt32("Close Window on Cancel");
             packet.ReadUInt32E<QuestFlags>("Quest Flags");
